util: Extract changes.github.PullRequestMixin to reusable module
